Create Exponentiation_II.cpp
Using fermants theorem
Prelude
We want compute the modular exponentiation A(BC) mod p = ?, where A, B, C, and p are known and p is a prim number. For example: 243mod 23 = 6

If we compute it in a straightforward way, first BC = e, and then Ae = n, and finally n mod p; we will run into the problem of creating (potentially) very large intermediary results for e and n. For example: e = 43 = 64, n = 264 ≈ 1.845x1019, and finally n mod 23 = 6

However, doing it in the straightforward way we did not take advantage of the fact that p is a prim number and that we are doing modular exponentiation. And also doing so, we will run into problems computing the result with a computer program in terms of time (CPU) and space (memory).

(Yes, we could do fast modular exponentiation using the identity (a ⋅ b) mod m = [(a mod m) ⋅ (b mod m)] mod m by first reducing A(BC) mod p to Ae mode p. For exmaple A2 mod p = (A ⋅ A) mod p = [(A mod p) ⋅ (A mod p)] mod p – but that is not where we want to go with this).

The smart way – use Fermat's little theorem
As documented in Find power of power under mod of a prime on GeeksforGeeks and the origin of this question, our exponent BC in A(BC) mod p can be expressed differently using Fermat's little theorem.

Fermat's little theorem states: a(p - 1) ≡ 1 (mod p) if p is a prime

Leading to following transformations:

It is possible to rewrite the exponent BC as x ⋅ (p - 1) + y

Using that alternate expression our ABC becomes Ax ⋅ (p - 1) + y = Ax ⋅ (p - 1) ⋅ Ay

Using Fermat's little theorem Ax ⋅ (p - 1) = 1; computing A(BC) mod p becomes computing Ay

Using BC = x ⋅ (p - 1) + y then y can be written as BC mod (p - 1)

From the above we get A(BC) mod p = (Ay) mod p

And with all of that we can compute A(BC) mod p in two steps while keeping the intermediary results small.

y = (BC) mod (p - 1)
result = (Ay) mod p
For example: 243mod 23

y = 43 mod (23 - 1) = 64 mod 22 = 20
result = 220 mod 23 = 1048576 mod 23 = 6