Create Removal_Game.cpp
This question made me learn a lot of concepts and the very first thing that how not being clear in the beginning makes us waste a lot of time.

There is a list of n numbers and two players who move alternately. On each move, a player removes either the first or last number from the list, and their score increases by that number. Both players try to maximize their scores.
What is the maximum possible score for the first player when both players play optimally?
Input
The first input line contains an integer n: the size of the list.
The next line has n integers x1,x2,…,xn: the contents of the list.
Output
Print the maximum possible score for the first player.