Create Concert_Tickets.cpp
// Application for multiset and use of upperbound
For each customer, we want to find the maximum possible ticket price they will accept from the available tickets. If the customer's maximum price is too low, so that no ticket price is below it, we can return -1. If not, we find the greatest ticket price less than or equal to the maximum price, and then remove that price as specified by the problem. We use a multiset for this.